requirements.txt
transformers
PyMuPDF
googlesearch-python
beautifulsoup4
requests
Flask
torch
nltk